Pre-populated database info should be in UTC.  - Also making times more realistic rather than generated whatever times
This `ClockImpl` should return a `DateTime` in UTC timezone

`DateTimeCreatorHelper` should also return a `DateTime` in Timezone UTC

This service should not be setting the status on behalf of the app.  This was actually @hlbraddocks suggestion to remove it.. I hesitated and it burned me on a test.  Should have listened to @hlbraddock in the first place

Shore up service layer tests to do all checking in UTC and account for the updating of predefined times in the test database

These serializers and deserializers should also be set to write and read times with respect to UTC

Add all service tests to a group called 'dbtest', and make it so those are run by default during the mvn test phase.
 - If the database is clean to begin with, theses tests can be run and re-run w/o worrying about failures.  They clean up after themselves.

Shoring up `RegistratinUpdateProcessTests`